refacor: move shared types to app.tsx refactor: move fetch posts functionality to app.tsx
